Backend/AI/requirements.txt
smolagents
liteLLM
dotenv
# transformers
google-generativeai